pyproject.toml
[project]
name = "enkryptai-skill-scanner"
version = "0.1.0"
description = "Scan Agent Skill packages for security threats using multi-agent AI analysis"
authors = [{ name = "Enkrypt AI" }]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10,<3.14"
keywords = ["security", "agent", "skill", "scanner", "crewai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "crewai[tools]>=1.9,<2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
enkryptai-skill-scanner = "skill_scanner.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_scanner"]

# Ensure non-Python data files are included in the wheel
[tool.hatch.build.targets.wheel.force-include]
"src/skill_scanner/config" = "skill_scanner/config"
"src/skill_scanner/data" = "skill_scanner/data"
